Refactor to support apollo
